implemented History.java, made modifications to HistoryInterface.java

+++ src/exercise3/HistoryInterface.java
@@ -3,7 +3,7 @@ package exercise3;
 public interface HistoryInterface {
 	
 	public void clear();
-	public int get(int index);
-	public int add(int result);
-	public int remove(int index);
+	public double get(int index);
+	public double add(double result);
+	public double remove(int index);
 }
